Responsive Images: do it for reals!
Now, which image url is used can be adjusted after DOM creation. If
utils.adjustResponsiveImage(img), a new src url will be chosed based on the
current width of the image. Also, we memoize the last recorded image size
so that upon creation of a new one of the same type we can start it out at the
same size as other images. This prevents double-downloading if the initialWidth
is small and the current width is large.