Responsive Images: option to use parent width
Sometimes it makes sense for an image to fill its container instead of
using its current width as a guide. This was noticable when resizing
the browser fro very small (guide image was 'medium') to very large.
The image expanded to its natural dimensions and then *that* width
was used to choose a different size, even if *that* width was smaller
than the container.

This also adjusts the size of invisible images so they have the
correct source upon becoming visible.